requirements.txt
streamlit
pandas
numpy
plotly
statsmodels
websocket-client
